requirements.txt
python-binance
python-dotenv
streamlit
